+++ exporter/prometheusexporter/README.md
@@ -16,7 +16,7 @@ The following settings can be optionally configured:
 - `namespace` (no default): if set, exports metrics under the provided value.
 - `send_timestamps` (default = `false`): if true, sends the timestamp of the underlying
   metric sample in the response.
-- `metric_expiration` (default = `120m`): defines how long metrics are exposed without updates
+- `metric_expiration` (default = `5m`): defines how long metrics are exposed without updates
 
 Example:
 
@@ -31,6 +31,3 @@ exporters:
     send_timestamps: true
     metric_expiration: 180m
 ```
-
-The full list of settings exposed for this exporter are documented [here](./config.go)
-with detailed sample configurations [here](./testdata/config.yaml).
